Update tests directory and add a load_conf test
